requirements.txt
fastapi
pandas
tensorflow
haversine
uvicorn[standard]
mysql-connector-python